requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.1.0
attrs==19.3.0
Automat==0.8.0
av==12.0.0
backcall==0.2.0
blinker==1.4
cachetools==5.2.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
distro-info===0.23ubuntu1
docker-pycreds==0.4.0
einops==0.7.0
entrypoints==0.3
executing==1.2.0
fasteners==0.19
filelock==3.15.4
fonttools==4.38.0
fsspec==2024.6.1
gitdb==4.0.9
GitPython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==1.0.0
grpcio==1.51.0
h5py==3.9.0
httplib2==0.14.0
huggingface-hub==0.23.4
hyperlink==19.0.0
idna==2.8
ifxdaq @ file:///mnt/c/Users/HanaL/Documents/RadarFusionGUI/ifxdaq/ifxdaq-4.0.0-py3-none-any.whl
ifxradarsdk @ file:///mnt/c/Users/HanaL/Documents/RadarFusionGUI/radar_sdk/python_wheels/ifxradarsdk-3.5.0%2B8c595dbb-py3-none-linux_x86_64.whl
importlib-metadata==5.0.0
incremental==16.10.1
ipython==8.6.0
jedi==0.18.2
Jinja2==2.10.1
joblib==1.4.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
more-itertools==4.2.0
netifaces==0.10.4
numcodecs==0.12.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
omegaconf==2.3.0
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrealsense2==2.55.1.6486
pyrsistent==0.15.5
pyserial==3.4
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.6
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.9.3
seaborn==0.12.1
SecretStorage==2.3.1
sentry-sdk==2.7.1
service-identity==18.1.0
setproctitle==1.3.3
shiboken6==6.6.3.1
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.3
ssh-import-id==5.10
stack-data==0.6.1
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==0.9.10
torch==1.13.0
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.5.0
Twisted==18.9.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.19
wadllib==1.3.3
wandb==0.15.7
wcwidth==0.2.5
Werkzeug==2.2.2
zarr==2.16.1
zipp==1.0.0
zope.interface==4.7.1
